+++ psyplot/__init__.py
@@ -123,24 +123,39 @@ def get_versions(requirements=True, key=None):
             }
         }
     """
-    from pkg_resources import iter_entry_points
+    from importlib.metadata import entry_points
     ret = {'psyplot': _get_versions(requirements)}
-    for ep in iter_entry_points(group='psyplot', name='plugin'):
+
+    try:
+        eps = entry_points(group='psyplot', name='plugin')
+    except TypeError:  # python<3.10
+        eps = [ep for ep in entry_points().get('psyplot', [])
+                if ep.name ==  'plugin']
+    for ep in eps:
         if str(ep) in rcParams._plugins:
             logger.debug('Loading entrypoint %s', ep)
-            if key is not None and not key(ep.module_name):
+
+            try:
+                ep.module
+            except AttributeError:  # python<3.10
+                ep.module = ep.pattern.match(ep.value).group("module")
+
+            if key is not None and not key(ep.module):
                 continue
             try:
                 mod = ep.load()
             except (ImportError, ModuleNotFoundError) as e:
-                logger.debug("Could not import %s" % ep, exc_info=True)
-                logger.warning("Could not import %s" % ep)
-            try:
-                ret[str(ep.module_name)] = mod.get_versions(requirements)
-            except AttributeError:
-                ret[str(ep.module_name)] = {
-                    'version': getattr(mod, 'plugin_version',
-                                       getattr(mod, '__version__', ''))}
+                logger.debug("Could not import %s" % (ep, ), exc_info=True)
+                logger.warning("Could not import %s" % (ep, ), exc_info=True)
+            else:
+                try:
+                    ret[str(ep.module)] = mod.get_versions(requirements)
+                except AttributeError:
+                    ret[str(ep.module)] = {
+                        'version': getattr(
+                            mod, 'plugin_version',
+                            getattr(mod, '__version__', ''))
+                        }
     if key is None:
         try:
             import psyplot_gui